% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rnames}
\alias{print.rnames}
\title{A print method for rnames}
\usage{
\method{print}{rnames}(x, ...)
}
\arguments{
\item{x}{A rnames object}

\item{...}{Undocumented}
}
\value{
No return, custom print method for rnames objects.
}
\description{
A customized printed display for rnames output
}
