% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnames.R
\name{rnames.list}
\alias{rnames.list}
\title{General rnames method for lisis}
\usage{
\method{rnames}{list}(obj = obj, ignore = c(), ...)
}
\arguments{
\item{obj}{A list}

\item{ignore}{A set of sublists to be ignored}

\item{...}{Undocumented}
}
\value{
A list with rname class.
}
\description{
General rnames method for generic lists.
}
